gcp-microservices-demo/src/emailservice/emailservice/requirements.txt
backoff==1.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Deprecated==1.2.13
googleapis-common-protos==1.56.2
grpcio==1.46.3
idna==3.3
opentelemetry-api==1.11.1
opentelemetry-exporter-otlp-proto-grpc==1.11.1
opentelemetry-proto==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
protobuf==3.20.1
requests==2.28.0
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1